requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-json-field==0.5.7
django-jsonfield==1.4.1
gunicorn==20.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
